refactor: update the milestone-exists script to use kd_clients
