+++ .vscode/launch.json
@@ -163,6 +163,7 @@
                 "--allow-all"
             ],
             "args": [
+				"KinsonDigital",
 				"Infrastructure",
 				"v1.2.3-preview.4",
 				"${env:CICD_TOKEN}",
